[refactor/#43] SearchBar 접근성 코드 정리
